[Refactor] Update AggStateDesc to include result nullable flag and adjust related usages
