--- fe/README
@@ -1,14 +1,0 @@
-# fe-testing
-
-This module is used to store some common classes of other modules.
-
-# spark-dpp
-
-This module is Spark DPP program, used for Spark Load function.
-Depends: fe-testing
-
-# fe-core
-
-This module is the main process module of FE.
-Depends: fe-testing, spark-dpp
-
