+++ fe/fe-core/src/main/java/com/starrocks/catalog/Function.java
@@ -40,7 +40,6 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.gson.annotations.SerializedName;
-import com.starrocks.catalog.combinator.AggStateDesc;
 import com.starrocks.common.Pair;
 import com.starrocks.common.io.Writable;
 import com.starrocks.sql.ast.HdfsURI;
@@ -50,6 +49,7 @@ import com.starrocks.sql.common.TypeManager;
 import com.starrocks.thrift.TFunction;
 import com.starrocks.thrift.TFunctionBinaryType;
 import com.starrocks.thrift.TTypeDesc;
+import com.starrocks.type.AggStateDesc;
 import com.starrocks.type.InvalidType;
 import com.starrocks.type.Type;
 import com.starrocks.type.TypeSerializer;
@@ -765,7 +765,7 @@ public class Function implements Writable {
             fn.setChecksum(checksum);
         }
         if (aggStateDesc != null) {
-            fn.setAgg_state_desc(aggStateDesc.toThrift());
+            fn.setAgg_state_desc(TypeSerializer.toThrift(aggStateDesc));
         }
         fn.setCould_apply_dict_optimize(couldApplyDictOptimize);
         return fn;
